backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
psycopg2-binary
sqlalchemy==2.0.23
asyncpg
qdrant-client==1.7.0
google-generativeai==0.4.1
python-dotenv==1.0.0
tiktoken==0.5.2
langchain==0.0.339
langchain-community==0.0.28
langchain-google-genai==0.0.11
python-multipart==0.0.6
cryptography
httpx==0.25.2